requirements.txt
# Google ADK for Python
google-adk>=1.12.0

# Notion integration
notion-client>=2.2.15

# GitHub integration
PyGithub>=2.3.0

# AI & Voice
google-generativeai>=0.8.0
elevenlabs>=1.0.0
openai>=1.54.0  # For Whisper

# PDF parsing
PyPDF2>=3.0.0
pdfminer.six>=20221105

# Utilities
python-dotenv>=1.0.0
requests>=2.31.0

